Add more unit-level test for CompleteTodo use case
As part of this, added some useful helpers to TestUtils and finally
convinced myself of what I was pretty sure of from the start, which was
that using checked exceptions wasn't a good idea after all. The deciding
factor was that they prevent use cases from being used within lambdas.