Move outcome definitions to top of use case classes
The motivation here is that the most interesting part of the use case is
what it accomplishes, and the outcome interface provides a really nice,
noiseless summary of the things that could happen when you invoke the
use case. So, having it at the top of the file makes for a better
opening-a-use-case-file experience.