Create givenRecordExists helper for use cases that perform lookups
This has to do with a subtle behavior of Optionals. In the following
expression:

```java
return someOptional
  .map(this::doWork)
  .orElseGet(this::explode);
```

`this.explode` will be invoked in a number of scenarios:

1. If the optional is empty
2. If `this.doWork` returns null

That second scenario poses a problem for "void" use cases that rely on
side effects of the outcome methods, which so far I've accomplished by
returning null in each outcome method.

The givenRecordExists helper performs Optional-like work, but doesn't
"help" in the case that the handler function returns null.

I was conflicted about how to test this. The bug arises across a lot of
use cases, but I couldn't find a way to generalize the test to
facilitate applying it everywhere it was relevant. So in this commit,
there's no test coverage that forces the use cases to use
givenRecordExists, which spooks me.

I may also opt for a homebrew Optional implementation that doesn't have
this behavior, which would look more familiar. Perhaps a "strict"
function that wraps an Optional and makes it less friendly.