Implement SpawnRecurringTask use case
Includes entities and repository for RecurringTasks

Also created StrictOptional, as a nicer solution than
`givenRecordExists` for the outcomes-returning-null problem.

One interesting assumption this implementation makes is that the use
case will be run on a very frequent interval. The logic checks to see
whether the most recent non-future scheduled occurrence of the recurring task
(if any) has an instance of the task created on or after it. If not, it
creates an instance of the task _at the current time._

This is the only time the task could be created, of course. But it
does introduce a scenario where the latest scheduled occurrence was two
days ago, and for some reason the spawning job hasn't run in all that
time. This would result in the task instance seeming to spawn two days
late.