Fix bug in gherkin runner and RecurringTaskSubForm
Some trickery with Vue.nextTick() was causing the tests to pass even
though changes to the schedule parameter (for example, everyXDays)
weren't triggering event emission.

In reproducing the issue with a test, I found that the way the gherkin
runner ran the Given and When steps didn't handle asynchronous steps
well; making asynchronous given and when steps run properly in serial
was necessary to drive out the event emission problem.