Move modules into a packages directory
This is more idiomatic with other projects, so it'll be more familiar to
others, and it removes the need to modify the root package.json's
workspaces list when a new package is added.

(Trade off: you have to dig one level below the root to see what things
there are. So the architecture screams, but only once you open the right
door?)