Extract gherkin-runner to separate module
Discovered while doing this some weirdness involving babel
transpilation. gherkin-runner wasn't getting transpiled when tests run.

Found some discussion of the issue, including an indication that a fix
is due out in jest soon:

https://github.com/facebook/jest/pull/4761

I updated jest to 21.3.0-beta.3 to see if that helped. It didn't at
first, which turned out to be because it wasn't under the original
package directory, so babel didn't know what transforms to run (?).
(These thoughts are based on discussion here:)

https://github.com/babel/babel-loader/issues/293

I added some babel configuration to the root-level package.json which
fixed the issue.